Block editor: extract iframe component
